move module order detail screen
