+++ ui/src/main/java/com/jin/ui/order/detail/OrderDetailScreen.kt
@@ -1,4 +1,4 @@
-package com.jin.honey.feature.orderdetail.ui
+package com.jin.ui.order.detail
 
 import android.widget.Toast
 import androidx.compose.foundation.background
@@ -35,22 +35,22 @@ import androidx.compose.ui.unit.dp
 import com.jin.feature.ui.state.DbState
 import com.jin.feature.ui.state.SearchState
 import com.jin.feature.ui.state.UiState
-import com.jin.honey.R
+import com.jin.ui.R
 import com.jin.ui.cart.content.CartOptionModifyBottomSheet
 import com.jin.ui.address.LocationSearchBottomSheet
 import com.jin.domain.order.model.Order
 import com.jin.domain.order.model.PayPrice
 import com.jin.domain.order.model.PaymentState
 import com.jin.domain.order.model.Requirement
-import com.jin.honey.feature.orderdetail.ui.content.NeedAgreeToTermsDialog
-import com.jin.honey.feature.orderdetail.ui.content.OrderAddress
-import com.jin.honey.feature.orderdetail.ui.content.OrderDetailAgreeToTerms
-import com.jin.honey.feature.orderdetail.ui.content.OrderDetailCartItems
-import com.jin.honey.feature.orderdetail.ui.content.OrderDetailHeader
-import com.jin.honey.feature.orderdetail.ui.content.OrderDetailPayment
-import com.jin.honey.feature.orderdetail.ui.content.OrderDetailPrice
-import com.jin.honey.feature.orderdetail.ui.content.OrderDetailRequirements
-import com.jin.honey.feature.orderdetail.ui.content.RiderRequirementBottomSheet
+import com.jin.ui.order.detail.content.NeedAgreeToTermsDialog
+import com.jin.ui.order.detail.content.OrderAddress
+import com.jin.ui.order.detail.content.OrderDetailAgreeToTerms
+import com.jin.ui.order.detail.content.OrderDetailCartItems
+import com.jin.ui.order.detail.content.OrderDetailHeader
+import com.jin.ui.order.detail.content.OrderDetailPayment
+import com.jin.ui.order.detail.content.OrderDetailPrice
+import com.jin.ui.order.detail.content.OrderDetailRequirements
+import com.jin.ui.order.detail.content.RiderRequirementBottomSheet
 import com.jin.domain.address.model.SearchAddress
 import com.jin.ui.theme.PointColor
 import java.text.NumberFormat
