+++ ui/src/main/java/com/jin/ui/order/detail/content/OrderDetailPrice.kt
@@ -1,4 +1,4 @@
-package com.jin.honey.feature.orderdetail.ui.content
+package com.jin.ui.order.detail.content
 
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
@@ -13,7 +13,7 @@ import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
-import com.jin.honey.R
+import com.jin.ui.R
 
 @Composable
 fun OrderDetailPrice(modifier: Modifier, productPrice: String, ridePrice: String, totalPrice: String) {
